block users form destroying different applicants

+++ app/controllers/user_infos_controller.rb
@@ -1,10 +1,13 @@
 class UserInfosController < ApplicationController
   def destroy
     user_info = UserInfo.find(params[:id])
-    applicant = Applicant.where(user: user_info.user.id, team: user_info.team.id).take
     
-    user_info.destroy!
-    applicant.destroy!
+    if user_info.user == current_user
+      applicant = Applicant.where(user: user_info.user.id, team: user_info.team.id).take
+
+      user_info.destroy!
+      applicant.destroy!
+    end
 
     redirect_to :back
   end
